@@ -2,7 +2,7 @@
     "name": "benjamincrozat/blade",
     "description": "Use Laravel Blade in any PHP project.",
     "homepage": "https://github.com/benjamincrozat/bade",
-    "license": "MIT",
+    "license": "WTFPL",
     "authors": [
         {
             "name": "Benjamin Crozat",
@@ -19,11 +19,11 @@
         "templates"
     ],
     "require": {
-        "php": "5.6.*|7.*",
-        "illuminate/view": "5.*|6.*|7.*"
+        "php": "^7.4|^8.0",
+        "illuminate/view": "^8.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "5.*|6.*|7.*"
+        "phpunit/phpunit": "^9.5"
     },
     "autoload": {
         "psr-4": {
